// src/app/(modulos)/operaciones/models/models.ts
export interface Categoria {
  id: number;
  value: string;
}

export interface Color {
  id: string;
  name: string;
  sku: string;
  hexadecimal: string;
  isActive: boolean;
}

export interface MecsaColor {
  id: string;
  name: string;
  sku: string;
  hexadecimal: string;
  isActive: boolean;
}

export interface MecsaColorResponse {
  mecsaColors: MecsaColor[];
}

export interface Fibra {
  id: string;
  categoryId: number;
  denomination: string | null;
  origin: string | null;
  colorId: string | null;
  isActive: boolean;
  category: Categoria;
  color: MecsaColor | null;
}

export interface FiberCategory {
  id: number;
  value: string;
}

export interface FibraResponse {
  fibers: Fiber[];
}

export interface Fiber {
  id: string;
  denomination: string;
  origin: string | null;
  isActive: boolean;
  category: Categoria;
  color: Color | null;
}